merge: frontend into develop

See merge request s06-final/S06P31S104!142

+++ front/bwadrigo/src/components/usePlan/SurveyTermination.jsx
@@ -2,54 +2,65 @@ import axios from "axios";
 import { useCallback, useState } from "react";
 import { Button } from "react-bootstrap";
 import surveyData from "../../surveyData.json";
+import styles from "./UsePlan.css";
 
 function SurveyTermination() {
-  const goFinish = useCallback(async(e) => {
-    let Authorization = localStorage.getItem("authorization")
-    let RefreshToekn = localStorage.getItem("refreshtoken")
+  const goFinish = useCallback(async (e) => {
+    if(reason == "") {
+      alert("사유를 선택해주세요.");
+      return;
+    }
+    let Authorization = localStorage.getItem("authorization");
+    let RefreshToekn = localStorage.getItem("refreshtoken");
     await axios({
-      method : "delete",
-      url : "/v1/api/plan",
-      data : reason,
-      headers : {
-        "Authorization" : Authorization,
-        "RefreshToken" : RefreshToekn 
-      }
-    }).then((res)=>{
-      console.log(res)
+      method: "delete",
+      url: "/v1/api/plan",
+      data: {reason: reason},
+      headers: {
+        Authorization: Authorization,
+        RefreshToken: RefreshToekn,
+      },
+    }).then((res) => {
+      console.log(res);
       window.location.href = "/finish-termination";
-    })
+    });
   });
-  const [reason, setReason] = useState("")
+  const [reason, setReason] = useState("");
   const goCancle = () => {
     window.location.href = "/member-plan";
   };
 
   const changeReason = (e) => {
     console.log(e.target.value);
-    setReason(e.target.value)
+    setReason(e.target.value);
   };
   return (
     <div>
-      <h3>서비스 해지</h3>
-      <div>
-        <b>서비스 해지 사유를 알려주세요.</b>
+      <h3 className="text-center Title">서비스 해지</h3>
+      <hr />
+      <div className="full-content">
+        <div>
+          <b style={{ fontSize: "20px", paddingLeft: "10px" }}>서비스 해지 사유를 알려주세요.</b>
+        </div>
+        <br />
+        <div style={{ paddingLeft: "20px" }}>
+          {surveyData.reasons.map((item) => (
+            <div key={item.id}>
+              <label style={{ paddingBottom: "10px" }}>
+                <input type="radio" name="survey" value={item.content} onChange={changeReason} />
+                <span> {item.content}</span>
+              </label>
+            </div>
+          ))}
+        </div>
       </div>
       <div>
-        {surveyData.reasons.map((item) => (
-          <div key={item.id}>
-            <input type="radio" name="survey" value={item.content} onChange={changeReason} />
-            <span> {item.content}</span>
-          </div>
-        ))}
-      </div>
-      <div>
-        <Button variant="primary" onClick={goCancle}>
+        <button className="btnPlan" onClick={goCancle}>
           해지 취소하기
-        </Button>
-        <Button variant="light" onClick={goFinish}>
+        </button>
+        <button className="btnCancle" onClick={goFinish}>
           즉시 해지하기
-        </Button>
+        </button>
       </div>
     </div>
   );
